gsysPortViewer.cc: Correct cast issue for 64-bit systems

+++ gsysPortViewer.cc
@@ -153,12 +153,12 @@ void gsysPortViewer::pushButton1_clicked()
       table1->setNumRows( table1->numRows() + 1 );
       table1->setText(table1->numRows()-1, 0, aktPort->getName());
       table1->setText(table1->numRows()-1, 1, aktPort->getValue());
-      table1->setText(table1->numRows()-1, 2, (new gsysMain())->getSimulator()->asChar((int)aktPort));
+      table1->setText(table1->numRows()-1, 2, (new gsysMain())->getSimulator()->asChar((int64_t)aktPort));
       portVec.erase(portVec.begin());
     }
     aktPort = 0;
   }
-  else		// anderen Port hinzufügen
+  else		// anderen Port hinzufÃ¼gen
   {
     int idx = comboBox1->currentItem();
     gsysPort* aktPort = portVec[idx-1];
@@ -166,7 +166,7 @@ void gsysPortViewer::pushButton1_clicked()
     table1->setNumRows( table1->numRows() + 1 );
     table1->setText(table1->numRows()-1, 0, aktPort->getName());
     table1->setText(table1->numRows()-1, 1, aktPort->getValue());
-    table1->setText(table1->numRows()-1, 2, (new gsysMain())->getSimulator()->asChar((int)aktPort));
+    table1->setText(table1->numRows()-1, 2, (new gsysMain())->getSimulator()->asChar((int64_t)aktPort));
     portVec.erase(portVec.begin()+idx-1);
   }
 }
